Issue #94: game session history - React side   - move all history pages to a single directory
