+++ src/routes.js
@@ -4,9 +4,9 @@ import Login from './components/LoginPages/login';
 import PlayerWaitinglists from './components/PlayerPages/playerWaitinglists';
 import GameWaitinglists from './components/GamePages/gameWaitinglists';
 import CardSelection from './components/GamePages/cardSelection';
-import GameHistory from './components/GamePages/gameHistory';
-import GameHistories from './components/GamePages/gameHistories'
-import PlayerGameHistories from './components/PlayerPages/playerGameHistories';
+import GameHistory from './components/HistoryPages/gameHistory';
+import GameHistories from './components/HistoryPages/gameHistories'
+import PlayerGameHistories from './components/HistoryPages/playerGameHistories';
 
 export const routes = [
   {
